Support multiple signaturePublickeys
Allow authentication servers to specify an array of profile property
public keys instead of just one.

Adds a new field, `signaturePublickeys`, to the authlib-injector
metadata route at the root of the API. `signaturePublickeys` is an
array of strings containing public keys in the same format as the
`signaturePublickey`. Any of the keys specified either in
`signaturePublickeys` or as the `signaturePublickey` can be used to
verify the signature of a textures payload.

With this change, authentication server A can forward a texture payload
from authentication server B, and the client could accept a texture
signed by either authentication server.